+++ wandb/data_types.py
@@ -33,10 +33,10 @@ _PY3 = sys.version_info.major == 3 and sys.version_info.minor >= 6
 
 if _PY3:
     from wandb.sdk.interface import _dtypes
-    from wandb.sdk.data_types import WBValue
+    from wandb.sdk.data_types import WBValue, Histogram
 else:
     from wandb.sdk_py27.interface import _dtypes
-    from wandb.sdk_py27.data_types import WBValue
+    from wandb.sdk_py27.data_types import WBValue, Histogram
 
 
 def _safe_sdk_import():
@@ -66,74 +66,6 @@ DATA_FRAMES_SUBDIR = os.path.join("media", "data_frames")
 
 def wb_filename(key, step, id, extension):
     return "{}_{}_{}{}".format(key, step, id, extension)
-
-
-class Histogram(WBValue):
-    """wandb class for histograms.
-
-    This object works just like numpy's histogram function
-    https://docs.scipy.org/doc/numpy/reference/generated/numpy.histogram.html
-
-    Examples:
-        Generate histogram from a sequence
-        ```
-        wandb.Histogram([1,2,3])
-        ```
-
-        Efficiently initialize from np.histogram.
-        ```
-        hist = np.histogram(data)
-        wandb.Histogram(np_histogram=hist)
-        ```
-
-    Arguments:
-        sequence: (array_like) input data for histogram
-        np_histogram: (numpy histogram) alternative input of a precoomputed histogram
-        num_bins: (int) Number of bins for the histogram.  The default number of bins
-            is 64.  The maximum number of bins is 512
-
-    Attributes:
-        bins: ([float]) edges of bins
-        histogram: ([int]) number of elements falling in each bin
-    """
-
-    MAX_LENGTH = 512
-
-    def __init__(self, sequence=None, np_histogram=None, num_bins=64):
-
-        if np_histogram:
-            if len(np_histogram) == 2:
-                self.histogram = (
-                    np_histogram[0].tolist()
-                    if hasattr(np_histogram[0], "tolist")
-                    else np_histogram[0]
-                )
-                self.bins = (
-                    np_histogram[1].tolist()
-                    if hasattr(np_histogram[1], "tolist")
-                    else np_histogram[1]
-                )
-            else:
-                raise ValueError(
-                    "Expected np_histogram to be a tuple of (values, bin_edges) or sequence to be specified"
-                )
-        else:
-            np = util.get_module(
-                "numpy", required="Auto creation of histograms requires numpy"
-            )
-
-            self.histogram, self.bins = np.histogram(sequence, bins=num_bins)
-            self.histogram = self.histogram.tolist()
-            self.bins = self.bins.tolist()
-        if len(self.histogram) > self.MAX_LENGTH:
-            raise ValueError(
-                "The maximum length of a histogram is %i" % self.MAX_LENGTH
-            )
-        if len(self.histogram) + 1 != len(self.bins):
-            raise ValueError("len(bins) must be len(histogram) + 1")
-
-    def to_json(self, run=None):
-        return {"_type": "histogram", "values": self.histogram, "bins": self.bins}
 
 
 class Media(WBValue):
@@ -2969,3 +2901,6 @@ class _TableType(_dtypes.Type):
 _dtypes.TypeRegistry.add(_ClassesIdType)
 _dtypes.TypeRegistry.add(_ImageType)
 _dtypes.TypeRegistry.add(_TableType)
+
+
+__all__ = ["WBValue", "Histogram"]
